Added joker icons, diamonds etc.

+++ app/src/main/java/com/kolip/numberle/MainActivity.java
@@ -1,13 +1,17 @@
 package com.kolip.numberle;
 
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
 public class MainActivity extends AppCompatActivity {
     GameManager gameManager;
+    GameStatus gameStatus;
+    GameStatus previousStatus;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -19,6 +23,7 @@ public class MainActivity extends AppCompatActivity {
 
     private void init() {
         gameManager = new GameManager(this, getBoxes(), getResultViews());
+        gameStatus = GameStatus.READY;
     }
 
     private ResultView[] getResultViews() {
@@ -70,12 +75,44 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void onKeyClicked(View view) {
+        if (gameStatus == GameStatus.FINISHED || gameStatus == GameStatus.BEFORE_FINISHED ||
+                gameStatus == GameStatus.IN_PROGRESS) {
+            return;
+        }
+
         if (view.getId() == R.id.key_Enter) {
+            previousStatus = gameStatus;
+            gameStatus = GameStatus.IN_PROGRESS;
             gameManager.enter();
         } else if (view.getId() == R.id.key_delete) {
             gameManager.delete();
         } else {
             gameManager.write(String.valueOf(((TextView) view).getText()));
+        }
+    }
+
+    public void gameFinished(NumberResult numberResult) {
+        setStatus(numberResult);
+        if (gameStatus == GameStatus.BEFORE_FINISHED) {
+            AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(this);
+            alertDialogBuilder.setMessage("Before Finished dialog");
+            alertDialogBuilder.create().show();
+        }
+        Log.d("GameFinished", "Game finished dialog has been finished.");
+
+    }
+
+    private void setStatus(NumberResult numberResult) {
+        gameStatus = previousStatus;
+
+        if (numberResult.getCorrectPositionNumberCount() == 4) {
+            gameStatus = GameStatus.FINISHED;
+        } else if (gameManager.getCurrentRow() == 6) {
+            gameStatus = GameStatus.BEFORE_FINISHED;
+        } else if (gameStatus == GameStatus.SECOND_CHANGE) {
+            gameStatus = GameStatus.FINISHED;
+        } else {
+            gameStatus = GameStatus.PLAYING;
         }
     }
 }
